+++ uitests/time_clock.js
@@ -1,0 +1,11 @@
+let wd = require("wd");
+
+exports.testBasic = function (driver) {
+    return driver
+        .back().sleep(200)
+        .title().should.become("Regression Test")
+        .elementById('time_clock').click().sleep(200)
+        .title().should.become("Time Clock")
+        .back().sleep(200)
+        .title().should.become("Regression Test")
+}